alloy_eip7702
Module constants

EIP-7702 constants.

Constants

EIP7702_CLEARED_DELEGATION

The EIP-7702 delegation designator for a cleared delegation.

EIP7702_DELEGATION_DESIGNATOR

The EIP-7702 delegation designator.

EIP7702_TX_TYPE_ID

Identifier for EIP7702’s set code transaction.

MAGIC

Magic number used to calculate an EIP7702 authority.

PER_AUTH_BASE_COST

An additional gas cost per EIP7702 authorization list item.

PER_EMPTY_ACCOUNT_COST

A gas refund for EIP7702 transactions if the authority account already exists in the trie.

SECP256K1N_HALF

The order of the secp256k1 curve, divided by two. Signatures that should be checked according to EIP-2 should have an S value less than or equal to this.